requirements.txt
spotdl
openai
telebot
spotipy
argparse
python-dotenv